requirement.txt
llama-index
openai
pypdf
python-dotenv
streamlit
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
pinecone